made it with inner components
